Extract server config into a config file.
Makes it easier to set new ports, certs, etc.
